made something that works, click space
