+++ Assets/Scenes.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 1ae55e9b0afc4b646a3c7f21c5f4c501
+folderAsset: yes
+timeCreated: 1523658846
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
