+++ Assets/Scenes/test.unity.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 5fe9ba703f45bc942adaf08ae898c298
+timeCreated: 1523658847
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
