+++ Assets/Scripts/Carmechanics.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: c87f0523061da0f4eb1a557b00865a2b
+timeCreated: 1523658984
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
