--- Assets/carmechanics.cs
@@ -1,27 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class carmechanics : MonoBehaviour {
-
-	// Use this for initialization
-	void Start () {
-		
-	}
-	
-	// Update is called once per frame
-	void Update () {
-
-        if(Input.GetKeyDown("space")) {
-
-            print("space key was pressed");
-        }
-
-    }
-
-    private void FixedUpdate() {
-        GetComponent<Rigidbody2D>().AddForce(new Vector2(2,0));
-
-
-    }
-}
